MAINT: (finally) split implementation (_funcs_impl.py) and normalizations (_funcs.py)
